[Common] moved configs under python/configs/
